rag-pdf-langchain/requirements.txt
chromadb
gradio
gradio[mcp]
langchain
langchain-community
langchain-core
langchain-ollama
langchain-chroma
ollama
pdfplumber
python-dotenv
